Add biome::rest_api::actix::login module
Adds a module to hold the actix implementation of the login
route, currently behind the biome-credentials and rest-api-
actix feature flags.
